Organize symmetry packages and classes
Split core package into axis, cluster and core. Axis contains the
classes that calculate axes properties, cluster contains the classes
related to clustering chains into subunits, and core contains the
algorithms and classes to detect symmetry
