+++ biojava-structure-gui/src/main/java/org/biojava/nbio/structure/symmetry/jmolScript/JmolSymmetryScriptGeneratorPointGroup.java
@@ -23,10 +23,10 @@
  */
 package org.biojava.nbio.structure.symmetry.jmolScript;
 
+import org.biojava.nbio.structure.cluster.Subunits;
+import org.biojava.nbio.structure.symmetry.axis.RotationAxisAligner;
 import org.biojava.nbio.structure.symmetry.core.Rotation;
-import org.biojava.nbio.structure.symmetry.core.RotationAxisAligner;
 import org.biojava.nbio.structure.symmetry.core.RotationGroup;
-import org.biojava.nbio.structure.symmetry.core.Subunits;
 import org.biojava.nbio.structure.symmetry.geometry.Polyhedron;
 import org.jcolorbrewer.ColorBrewer;
