+++ biojava-structure/src/main/java/org/biojava/nbio/structure/cluster/ChainClusterer.java
@@ -18,11 +18,13 @@
  *      http://www.biojava.org/
  *
  */
-package org.biojava.nbio.structure.symmetry.core;
+package org.biojava.nbio.structure.cluster;
 
 import org.biojava.nbio.structure.Atom;
+import org.biojava.nbio.structure.symmetry.core.SequenceAlignmentCluster;
 
 import javax.vecmath.Point3d;
+
 import java.util.*;
 
 /**
