+++ biojava-structure/src/main/java/org/biojava/nbio/structure/cluster/ClusterMerger.java
@@ -18,9 +18,12 @@
  *      http://www.biojava.org/
  *
  */
-package org.biojava.nbio.structure.symmetry.core;
+package org.biojava.nbio.structure.cluster;
 
 import java.util.*;
+
+import org.biojava.nbio.structure.symmetry.core.QuatSymmetryParameters;
+import org.biojava.nbio.structure.symmetry.core.SequenceAlignmentCluster;
 
 /**
  * Merges clusters based on their sequence identity. This class does the actual
