+++ biojava-structure/src/main/java/org/biojava/nbio/structure/cluster/ClusterProteinChains.java
@@ -18,9 +18,11 @@
  *      http://www.biojava.org/
  *
  */
-package org.biojava.nbio.structure.symmetry.core;
+package org.biojava.nbio.structure.cluster;
 
 import org.biojava.nbio.structure.Structure;
+import org.biojava.nbio.structure.symmetry.core.QuatSymmetryParameters;
+import org.biojava.nbio.structure.symmetry.core.SequenceAlignmentCluster;
 
 import java.util.Collections;
 import java.util.List;
