+++ biojava-structure/src/main/java/org/biojava/nbio/structure/cluster/ProteinSequenceClusterer.java
@@ -18,10 +18,12 @@
  *      http://www.biojava.org/
  *
  */
-package org.biojava.nbio.structure.symmetry.core;
+package org.biojava.nbio.structure.cluster;
 
 import org.biojava.nbio.structure.Atom;
 import org.biojava.nbio.structure.Structure;
+import org.biojava.nbio.structure.symmetry.core.QuatSymmetryParameters;
+import org.biojava.nbio.structure.symmetry.core.SequenceAlignmentCluster;
 
 import java.util.*;
