+++ biojava-structure/src/main/java/org/biojava/nbio/structure/cluster/SubunitGraph.java
@@ -18,7 +18,7 @@
  *      http://www.biojava.org/
  *
  */
-package org.biojava.nbio.structure.symmetry.core;
+package org.biojava.nbio.structure.cluster;
 
 import javax.vecmath.Point3d;
