+++ biojava-structure/src/main/java/org/biojava/nbio/structure/cluster/Subunits.java
@@ -19,7 +19,7 @@
  *
  */
 
-package org.biojava.nbio.structure.symmetry.core;
+package org.biojava.nbio.structure.cluster;
 
 import org.biojava.nbio.structure.symmetry.geometry.MomentsOfInertia;
 import org.biojava.nbio.structure.symmetry.geometry.SuperPosition;
