+++ biojava-structure/src/main/java/org/biojava/nbio/structure/cluster/UniqueSequenceList.java
@@ -18,7 +18,7 @@
  *      http://www.biojava.org/
  *
  */
-package org.biojava.nbio.structure.symmetry.core;
+package org.biojava.nbio.structure.cluster;
 
 import org.biojava.nbio.structure.Atom;
 import org.biojava.nbio.structure.Group;
