+++ biojava-structure/src/main/java/org/biojava/nbio/structure/symmetry/analysis/ScanSymmetry.java
@@ -33,6 +33,7 @@ import org.biojava.nbio.structure.Structure;
 import org.biojava.nbio.structure.StructureException;
 import org.biojava.nbio.structure.StructureIO;
 import org.biojava.nbio.structure.align.util.AtomCache;
+import org.biojava.nbio.structure.cluster.Subunits;
 import org.biojava.nbio.structure.io.FileParsingParameters;
 import org.biojava.nbio.structure.io.mmcif.AllChemCompProvider;
 import org.biojava.nbio.structure.io.mmcif.ChemCompGroupFactory;
@@ -40,7 +41,6 @@ import org.biojava.nbio.structure.rcsb.GetRepresentatives;
 import org.biojava.nbio.structure.symmetry.core.QuatSymmetryDetector;
 import org.biojava.nbio.structure.symmetry.core.QuatSymmetryParameters;
 import org.biojava.nbio.structure.symmetry.core.QuatSymmetryResults;
-import org.biojava.nbio.structure.symmetry.core.Subunits;
 import org.biojava.nbio.structure.symmetry.misc.ProteinComplexSignature;
 import org.biojava.nbio.structure.symmetry.utils.BlastClustReader;
 import org.biojava.nbio.structure.xtal.SpaceGroup;
