+++ biojava-structure/src/main/java/org/biojava/nbio/structure/symmetry/axis/HelixAxisAligner.java
@@ -18,12 +18,17 @@
  *      http://www.biojava.org/
  *
  */
-package org.biojava.nbio.structure.symmetry.core;
+package org.biojava.nbio.structure.symmetry.axis;
 
+import org.biojava.nbio.structure.cluster.Subunits;
+import org.biojava.nbio.structure.symmetry.core.Helix;
+import org.biojava.nbio.structure.symmetry.core.HelixLayers;
+import org.biojava.nbio.structure.symmetry.core.QuatSymmetryResults;
 import org.biojava.nbio.structure.symmetry.geometry.MomentsOfInertia;
 import org.biojava.nbio.structure.symmetry.geometry.SuperPosition;
 
 import javax.vecmath.*;
+
 import java.util.*;
 
 public class HelixAxisAligner extends AxisAligner {
